src/requirements.txt
azure-storage-blob
boto3
fastapi
google-cloud-storage
python-dotenv
httpx
python-multipart
pydantic
uvicorn